+++ oshi-core-java25/src/main/java/oshi/software/os/linux/LinuxOperatingSystemFFM.java
@@ -9,18 +9,22 @@ import java.io.IOException;
 import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
 import java.nio.file.Files;
+import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.driver.linux.proc.AuxvFFM;
 import oshi.ffm.linux.LinuxLibcFunctions;
 import oshi.ffm.linux.UdevFunctions;
 import oshi.software.os.FileSystem;
 import oshi.software.os.NetworkParams;
 import oshi.software.os.OSProcess;
 import oshi.software.os.OSProcess.State;
+import oshi.util.ExecutingCommand;
 import oshi.util.ParseUtil;
+import oshi.util.driver.linux.proc.Auxv;
 import oshi.util.linux.ProcPath;
 
 /**
@@ -36,6 +40,45 @@ import oshi.util.linux.ProcPath;
 public class LinuxOperatingSystemFFM extends LinuxOperatingSystem {
 
     private static final Logger LOG = LoggerFactory.getLogger(LinuxOperatingSystemFFM.class);
+
+    private static final long USER_HZ;
+    private static final long PAGE_SIZE;
+    static {
+        Map<Integer, Long> auxv = AuxvFFM.queryAuxv();
+        long hz = auxv.getOrDefault(Auxv.AT_CLKTCK, 0L);
+        USER_HZ = hz > 0 ? hz : ParseUtil.parseLongOrDefault(ExecutingCommand.getFirstAnswer("getconf CLK_TCK"), 100L);
+        long pagesz = auxv.getOrDefault(Auxv.AT_PAGESZ, 0L);
+        PAGE_SIZE = pagesz > 0 ? pagesz
+                : ParseUtil.parseLongOrDefault(ExecutingCommand.getFirstAnswer("getconf PAGE_SIZE"), 4096L);
+    }
+
+    /**
+     * Gets Jiffies per second, useful for converting ticks to milliseconds and vice versa.
+     *
+     * @return Jiffies per second.
+     */
+    public static long hz() {
+        return USER_HZ;
+    }
+
+    /**
+     * Gets Page Size, for converting memory stats from pages to bytes.
+     *
+     * @return Page Size in bytes.
+     */
+    public static long pageSize() {
+        return PAGE_SIZE;
+    }
+
+    @Override
+    public long getHz() {
+        return USER_HZ;
+    }
+
+    @Override
+    public long getPageSize() {
+        return PAGE_SIZE;
+    }
 
     /**
      * Identifies if the udev library was successfully loaded and all symbols bound via FFM. Also respects the