+++ oshi-core/src/main/java/oshi/hardware/platform/linux/LinuxCentralProcessorJNA.java
@@ -22,6 +22,7 @@ import com.sun.jna.platform.linux.Udev.UdevListEntry;
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.linux.proc.AuxvJNA;
 import oshi.jna.platform.linux.LinuxLibc;
+import oshi.software.os.linux.LinuxOperatingSystemJNA;
 import oshi.util.FileUtil;
 import oshi.util.ParseUtil;
 import oshi.util.driver.linux.proc.Auxv;
@@ -33,6 +34,10 @@ import oshi.util.tuples.Quartet;
  */
 @ThreadSafe
 final class LinuxCentralProcessorJNA extends LinuxCentralProcessor {
+
+    LinuxCentralProcessorJNA() {
+        super(LinuxOperatingSystemJNA.hz());
+    }
 
     @Override
     protected Quartet<List<LogicalProcessor>, List<ProcessorCache>, Map<Integer, Integer>, Map<Integer, String>> readTopologyWithUdev() {