+++ oshi-core/src/main/java/oshi/hardware/platform/linux/LinuxHardwareAbstractionLayerJNA.java
@@ -7,11 +7,14 @@ package oshi.hardware.platform.linux;
 import java.util.List;
 
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.hardware.CentralProcessor;
+import oshi.hardware.GlobalMemory;
 import oshi.hardware.HWDiskStore;
 import oshi.hardware.LogicalVolumeGroup;
 import oshi.hardware.NetworkIF;
 import oshi.hardware.PowerSource;
 import oshi.hardware.UsbDevice;
+import oshi.software.os.linux.LinuxOperatingSystemJNA;
 
 /**
  * JNA-based Linux hardware abstraction layer. Extends {@link LinuxHardwareAbstractionLayer}, overriding methods as FFM
@@ -19,6 +22,16 @@ import oshi.hardware.UsbDevice;
  */
 @ThreadSafe
 public final class LinuxHardwareAbstractionLayerJNA extends LinuxHardwareAbstractionLayer {
+
+    @Override
+    public GlobalMemory createMemory() {
+        return new LinuxGlobalMemory(LinuxOperatingSystemJNA.pageSize());
+    }
+
+    @Override
+    public CentralProcessor createProcessor() {
+        return new LinuxCentralProcessorJNA();
+    }
 
     @Override
     public List<LogicalVolumeGroup> getLogicalVolumeGroups() {